pyproject.toml
[tool.poetry]
name = "LangChain Demo"
version = "1.0.0"
description = ""
authors = ["Horatiu Cristea <horatiu.cristea@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
typer = "^0.12.3"
python-dotenv = "^1.0.1"
langchain = "^0.1.16"
langchain-openai = "^0.1.3"
chromadb = "^0.4.24"
numexpr = "^2.10.0"


[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
mypy = "^1.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
